requirements.txt
flask
dash
pillow
textblob
jupyter
